refac : PR 반영
