+++ src/main/java/atwoz/atwoz/community/command/domain/selfintroduction/exception/InvalidSelfIntroductionContentException.java
@@ -1,0 +1,7 @@
+package atwoz.atwoz.community.command.domain.selfintroduction.exception;
+
+public class InvalidSelfIntroductionContentException extends RuntimeException {
+    public InvalidSelfIntroductionContentException() {
+        super("셀프 소개 글은 최소 30자 이상이어야 합니다.");
+    }
+}
